requirements.txt
biopython
numpy
Pillow
pyspark
matplotlib
numba
seaborn
#also run pip install fcgr-0.1-py3-none-any.whl